+++ src/utils/ray.cpp
@@ -5,7 +5,7 @@
 
 Ray::Ray(Tuple origin, Tuple direction) : origin(origin), direction(direction) {}
 
-Tuple Ray::position(float t) const {
+Tuple Ray::position(double t) const {
     return origin + direction * t;
 }
 
@@ -15,9 +15,9 @@ Ray Ray::transform(const Matrix& m) const {
 
 std::vector<Intersection> Ray::intersect(const std::shared_ptr<Shape> shape) const {
 
-    std::vector<float> ts = shape->intersect(*this);
+    std::vector<double> ts = shape->intersect(*this);
     std::vector<Intersection> res;
-    for(float t : ts) {
+    for(double t : ts) {
         res.push_back(Intersection(t, shape));
     }
     return res;
